[tune/structure] Move CLI files into subpackage
